AutoEncoder / SupervisedFeedForwardNeuralNetwork code dup removal
