+++ ml4j-nn-impl/src/main/java/org/ml4j/nn/unsupervised/AutoEncoderImpl.java
@@ -14,45 +14,27 @@
 
 package org.ml4j.nn.unsupervised;
 
-import org.ml4j.Matrix;
-import org.ml4j.nn.BackPropagation;
-import org.ml4j.nn.ForwardPropagation;
-import org.ml4j.nn.ForwardPropagationImpl;
-import org.ml4j.nn.axons.TrainableAxons;
-import org.ml4j.nn.costfunctions.CrossEntropyCostFunction;
-import org.ml4j.nn.layers.DirectedLayerActivation;
-import org.ml4j.nn.layers.DirectedLayerGradient;
+import org.ml4j.nn.FeedForwardNeuralNetworkBase;
 import org.ml4j.nn.layers.FeedForwardLayer;
 import org.ml4j.nn.neurons.NeuronsActivation;
-
-import org.ml4j.nn.neurons.NeuronsActivationFeatureOrientation;
-import org.ml4j.nn.synapses.DirectedSynapses;
-import org.ml4j.nn.synapses.DirectedSynapsesGradient;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
 /**
- * Default implementation of AutoEncoder consisting of 2 default FeedForwardLayers.
+ * Default implementation of AutoEncoder consisting of FeedForwardLayers.
  *
  * @author Michael Lavelle
  */
-public class AutoEncoderImpl implements AutoEncoder {
+public class AutoEncoderImpl extends 
+    FeedForwardNeuralNetworkBase<AutoEncoderContext, AutoEncoder> implements AutoEncoder {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(AutoEncoderImpl.class);
-  
-  private List<FeedForwardLayer<?, ?>> layers;
-  
   /**
    * Default serialization id.
    */
   private static final long serialVersionUID = 1L;
-
+  
+  private static final Logger LOGGER = LoggerFactory.getLogger(AutoEncoderImpl.class);
+ 
   /**
    * Constructor for a simple 2-layer AutoEncoder.
    * 
@@ -61,9 +43,7 @@ public class AutoEncoderImpl implements AutoEncoder {
    */
   public AutoEncoderImpl(FeedForwardLayer<?, ?> encodingLayer,
       FeedForwardLayer<?, ?> decodingLayer) {
-    this.layers = new ArrayList<FeedForwardLayer<?, ?>>();
-    this.layers.add(encodingLayer);
-    this.layers.add(decodingLayer);
+      super(encodingLayer, decodingLayer);
   }
   
   /**
@@ -72,116 +52,7 @@ public class AutoEncoderImpl implements AutoEncoder {
    * @param layers The layers
    */
   public AutoEncoderImpl(FeedForwardLayer<?, ?>... layers) {
-    this.layers = new ArrayList<FeedForwardLayer<?, ?>>();
-    this.layers.addAll(Arrays.asList(layers));
-  }
-
-  @Override
-  public void train(NeuronsActivation trainingDataActivations, AutoEncoderContext trainingContext) {
-
-    final int numberOfTrainingIterations = trainingContext.getTrainingIterations();
-    
-    LOGGER.info("Training AutoEncoderImpl...");
-
-    List<TrainableAxons<?, ?, ?>> trainableAxonsList = new ArrayList<>();
-    for (int layerIndex = 0; layerIndex < getNumberOfLayers(); layerIndex++) {
-
-      FeedForwardLayer<?, ?> layer = getLayer(layerIndex);
-      for (DirectedSynapses<?> synapses : layer.getSynapses()) {
-        if (synapses.getAxons() instanceof TrainableAxons) {
-          TrainableAxons<?, ?, ?> axons = (TrainableAxons<?, ?, ?>) synapses.getAxons();
-          trainableAxonsList.add(axons);
-        }
-      }
-    }
-
-    for (int i = 0; i < numberOfTrainingIterations; i++) {
-
-      // Forward propagate the trainingDataActivations through the entire AutoEncoder
-      ForwardPropagation forwardPropagation =
-          forwardPropagate(trainingDataActivations, trainingContext);
-      
-      // We're going to use the cross entropy cross function to train the AutoEncoder
-      final CrossEntropyCostFunction costFunction = new CrossEntropyCostFunction();
-      
-      // When using the cross entropy cross function, the deltas we backpropagate
-      // end up being the difference between the target activations ( which are the 
-      // same as the trainingDataActivations as this is an AutoEncoder), and the
-      // activations resulting from the forward propagation
-      Matrix deltasM = forwardPropagation.getOutputs().getActivations()
-          .sub(trainingDataActivations.getActivations());
-      
-      // The deltas we back propagate are in the transposed orientation to the inputs
-      NeuronsActivation deltas = new NeuronsActivation(deltasM.transpose(),
-          forwardPropagation.getOutputs().isBiasUnitIncluded(),
-          NeuronsActivationFeatureOrientation.ROWS_SPAN_FEATURE_SET);
-     
-      // Back propagate the deltas through the nework
-      BackPropagation backPropagation = forwardPropagation.backPropagate(deltas, trainingContext);
-
-      // Obtain the gradients of each set of Axons we wish to train - for this example it is
-      // all the Axons
-      List<Matrix> trainableAxonsGradients = new ArrayList<>();
-      List<DirectedLayerGradient> reversed = new ArrayList<>();
-      reversed.addAll(backPropagation.getDirectedLayerGradients());
-      Collections.reverse(reversed);
-      for (DirectedLayerGradient gradient : reversed) {
-        for (DirectedSynapsesGradient synapsesGradient : gradient.getSynapsesGradients()) {
-          Matrix trainableAxonsGradient = synapsesGradient.getTrainableAxonsGradient();
-          if (trainableAxonsGradient != null) {
-            trainableAxonsGradients.add(synapsesGradient.getTrainableAxonsGradient());
-          }
-          
-        }
-      }
-      Collections.reverse(trainableAxonsGradients);
-      
-      double learningRate = trainingContext.getTrainingLearningRate();
-      
-      for (int axonsIndex = 0; axonsIndex < trainableAxonsGradients.size(); axonsIndex++) {
-        TrainableAxons<?, ?, ?> axons = trainableAxonsList.get(axonsIndex);
-        // Transpose the axon gradients into matrices that correspond to the orientation of the
-        // connection weights ( COLUMNS_SPAN_FEATURE_SET )
-        Matrix axonsGrad = trainableAxonsGradients.get(axonsIndex).transpose();
-        Matrix existingAxonsWeights = axons.getDetachedConnectionWeights();
-
-        // Adjust the weights of each set of Axons by subtracting the learning-rate scaled
-        // gradient matrices
-        Matrix newAxonsWeights = existingAxonsWeights.sub(axonsGrad.mul(learningRate));
-        axons.setConnectionWeights(newAxonsWeights);
-      }
-
-      // Obtain the cost from the cost function
-      double cost = costFunction.getCost(trainingDataActivations.getActivations(),
-          forwardPropagation.getOutputs().getActivations());
-      LOGGER.info("COST:" + cost);
-    }
-  }
-
-  @Override
-  public List<FeedForwardLayer<?, ?>> getLayers() {
-    return layers;
-  }
-
-  @Override
-  public int getNumberOfLayers() {
-    return layers.size();
-  }
-
-  @Override
-  public FeedForwardLayer<?, ?> getLayer(int layerIndex) {
-    return layers.get(layerIndex);
-  }
-
-  @Override
-  public FeedForwardLayer<?, ?> getFirstLayer() {
-    return layers.get(0);
-  }
-
-  @Override
-  public FeedForwardLayer<?, ?> getFinalLayer() {
-    return layers.get(getNumberOfLayers() - 1);
-
+      super(layers);
   }
 
   @Override
@@ -211,31 +82,7 @@ public class AutoEncoderImpl implements AutoEncoder {
   }
 
   @Override
-  public ForwardPropagation forwardPropagate(NeuronsActivation inputActivation,
-      AutoEncoderContext context) {
-    
-    int endLayerIndex =
-        context.getEndLayerIndex() == null ? (getNumberOfLayers() - 1) : context.getEndLayerIndex();
-
-    LOGGER.debug("Forward propagating through AutoEncoderMock from layerIndex:"
-        + context.getStartLayerIndex() + " to layerIndex:" + endLayerIndex);
-        
-    NeuronsActivation inFlightActivations = inputActivation;
-    int layerIndex = 0;
-    List<DirectedLayerActivation> activations = new ArrayList<>();
-    for (FeedForwardLayer<?, ?> layer : getLayers()) {
-
-      if (layerIndex >= context.getStartLayerIndex() && layerIndex <= endLayerIndex) {
-
-        DirectedLayerActivation inFlightLayerActivations = 
-            layer.forwardPropagate(inFlightActivations, context.createLayerContext(layerIndex));
-        activations.add(inFlightLayerActivations);
-        inFlightActivations = inFlightLayerActivations.getOutput();
-      }
-      layerIndex++;
-
-    }
-    
-    return new ForwardPropagationImpl(activations, inFlightActivations);
+  public void train(NeuronsActivation inputActivations, AutoEncoderContext trainingContext) {
+    super.train(inputActivations, inputActivations, trainingContext);
   }
 }
